# server: skip unnecessary sprintf which executes proto.Size()