From the profiling, I saw that when we call proto.Size(), it will unmarshal the data and return the size, which is costly. Similar to what we got in https://github.com/etcd-io/etcd/pull/12871

To avoid unnecessary proto.Size() call, I created this patch.

Some initial evaluation looks good. More data generating is in progress. It can increase the throughput by around 10% when we have small data size and a large number of requests.

@gyuho @ptabor