@nekto0n 2021-06-03
I thought maybe it is worth adding some benchmarks? Something as simple as:
```
func BenchmarkWarnOfExpensiveRequestNoLog(b *testing.B) {
	m := &raftpb.Message{
		Type:       0,
		To:         0,
		From:       1,
		Term:       2,
		LogTerm:    3,
		Index:      0,
		Entries:    []raftpb.Entry{
			{
				Term:  0,
				Index: 0,
				Type:  0,
				Data:  make([]byte, 1024),
			},
		},
		Commit:     0,
		Snapshot:   raftpb.Snapshot{},
		Reject:     false,
		RejectHint: 0,
		Context:    nil,
	}
	err := errors.New("some error text")
	for n := 0; n < b.N; n++ {
		warnOfExpensiveRequest(testLogger, time.Second, time.Now(), nil, m, err)
	}
}
```
Gives:
```
benchcmp old.txt new.txt 
benchmark                                  old ns/op     new ns/op     delta
BenchmarkWarnOfExpensiveRequestNoLog-8     1005          69.3          -93.11%
```
and
```
benchstat old.txt new.txt 
name                           old time/op  new time/op  delta
WarnOfExpensiveRequestNoLog-8  1.00µs ± 0%  0.07µs ± 0%   ~     (p=1.000 n=1+1)
```