@wilsonwang371 2021-06-03
This is the performance evaluation. We can see the best performance improvement can be around 10%. The worst case is around 4% degradation but I think this is not related to the new changes. and is within the performance variance.

![image](https://user-images.githubusercontent.com/3913185/120689510-0a5e0780-c459-11eb-87e2-e9dd57757f9b.png)