@wilsonwang371 2021-06-03
> @wilsonwang371 Let's add that benchmark case for future reference, otherwise LGTM. Thanks.

Just added the benchmark code.